+++ Makefile
@@ -1,8 +1,8 @@
 # $PostgreSQL: pgsql/contrib/plpgsql_check/Makefile
 
 MODULE_big = plpgsql_check
-OBJS = format.o catalog.o tablefunc.o plpgsql_setup.o expr_walk.o assign.o check_expr.o report.o stmtwalk.o
-DATA =  plpgsql_check--1.0--1.1.sql plpgsql_check--1.4.sql plpgsql_check--1.1--1.2.sql plpgsql_check--1.2--1.3.sql plpgsql_check--1.3--1.4.sql
+OBJS = format.o catalog.o tablefunc.o expr_walk.o assign.o check_expr.o report.o stmtwalk.o check_function.o typdesc.o profiler.o plpgsql_check.o
+DATA = plpgsql_check--1.5.sql
 EXTENSION = plpgsql_check
 
 ifndef MAJORVERSION
@@ -27,4 +27,4 @@ ifeq ($(PORTNAME), darwin)
 override CFLAGS += -undefined dynamic_lookup
 endif
 
-override CFLAGS += -I$(top_builddir)/src/pl/plpgsql/src
+override CFLAGS += -I$(top_builddir)/src/pl/plpgsql/src -Wall
