+++ plpgsql_check--1.5.sql
